Refactor PersistentConversationHandler and add usePersistentChatHistory hook for improved chat history management
